refactor: move spawn methods to own manager class
